rename: searchbar 폴더 input으로 이름 변경
