+++ src/components/inputs/SearchbarStyle.js
